app/fastapi/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
boto3==1.34.58
botocore==1.34.58
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-sessions==0.3.2
flatbuffers==24.3.6
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
jmespath==1.0.1
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-multipart==0.0.9
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.10.0
urllib3==2.0.7
uvicorn==0.27.1
Werkzeug==3.0.1
wrapt==1.14.1
